move documentation generation stuff to module
